Directory and Package Restructured.
The directory restructure has been finished, and packages have been fixed in the classes.

Minor changes have been made to some classes in order to make them able to compile properly.
Boilerplate code has been added to the top of every file - this is recommended for every new file from now on.

There is a severe error in the World class, which is causing runtime crashes when initialised.