World Behaviour changed. Platform and Background GFX changed.
##World Behaviour##
The world now loads more information from the xml file - such as the length and height of the level. It now allows you to move left and right in the level until near the ends - at which point it changes movement to a fixed screen and player moving to the edge.
There are issues with player positionition - this NEEDS fixed urgently; and will do after lunch.

##Graphics Changes##
Platform graphics and background have been changed to make them look better and less rubbish. Please do more work on these in the future - they are still placeholder graphics.