World Position Changed. Player start and end working.
##World Position##
The world positions itself to the bottom left of the screen as 0,0. This is to make it much easier to work with heights on the screen for Martin's level designs.
It loads in the data as -height on the XML file now - so he can add every height as if it was cartesian (x,y)

##Player Start##
The player starts near the left side, and can move till the end of the level. This is working much better than before.
ToDo: the player start should be in the xml file, instead of hardcoded.