Platform Colours in Level File
## Platform Colours##
The platform colours have now been changed to be relative to a base colour - instead of needing two separate colours. This means the top will always be darker.
This is a temporary class, but the change may stay for a while untill we have actual art.

##Level File##
The platform colour attibute <ob> <colour> HEX COLOUR </colour> .... is now necessary and the game may fail to load without it.
The world object should be updated to check for the colour given and ensure the entire file will not fail because of a bad input.