Platform top changed. PlayerEvent added.
##Platform Changes##
The top of the platform is now a fixed height, allowing for more versatile paltforms, such as the columns in the thursday example (long height, thin width, y that allows it to touch the floor).
This change should be superceeded by changes to allow different platform types, and a specified top height.
This will become redundant once art is implemented to replace it.

##Player Event##
This class is under work.
It is designed to be dispatched whenever the player is related to an action - such as getting stuck in jam, or getting hit in a collision.
The scope of this event should be limited to the player only - although similar events can be fired at the same time as these - that allow for loggers etc.

##Level XML##
Now holds changes that allow columns.
Also has some comments to make it easier to navigate.
This will all be devolved once the level editor is written.