Overlay Created
##Overlay Class##
The overlay class stands infront of the main gameplay, and holds the UI etc.
It's currently the same mechanic as the Background and so may be suited to merge the two classes into a single generic one.
Currently it holds temporary graphics for the prototype, but it will be extended to hold the outputs of the score, timers etc.