Player Throw Caught
### Player Throw ###
The player throw event is now caught in the world, and displays the properties correctly. The next steps are to add functionality to actually throw the physical object from the world.
Some typos have been cleaned up as well.

### Player Target ###
The player target has been cut down, prior to full-removal. It now displays the reticule for debugging purposes, but this will be relegated to other parts of the game in future.
All throwing functionality has been successfully moved to the world and Player now.