Throwing Event Catching
### Throw Event ###
The world now spawns an object to show the angle and magnitude of the throw. There are still some serious issues with this as the angle is never correct.
Please look into fixing either the Player.getAimingAngle(), or the world's onThrow listener.
Bug is expected to be in the getAimingAngle, at the point it converts the player.x and y to stage x and y.