Throwing Fixed
### Player Target ###
Now uses a parent reference for the event x and y, so as to allow any object to parent this and use it as their aiming system.
More updates to this should make it more generic and less visual; with an optional visual overlay - so enemies do not draw the same throwing stuff as the player.

### Player ###
Throwing has yet again been devolved from the player, it is now back to being in the player target, so this targetting can be reused repeatedly by enemies etc.

### World ###
The world now handles the trigonometry needed properly now, it puts the representative dot at the correct angle and magnitude from the player location.
This should now be made into a new Throwable object type, that will traverse it's trajectory untill a collision or some timer has run out.