Throwable Objects Working
### Throwable Objects ###
Throwable objects as a base class now function correctly. They move in a parabola; either in the x or y field - or using a mix of each.
For a straight line, the y and x accel should be 0.
For a standard gravitational parabola, use a yAccel and no xAccel
Other combinations can result in new styles of movement.

Currently lacks any collision detection, but that should be left for a specific sub class - these throwables are just the basis of how to movement object.