TestPlayer Collisions Prototype
### Test Player ###
The TestPlayer now collides in the expected way for these tests. It will jump to the top of the object when it hits is, and then stay still without shaking.
The movement involved in this, is simple and allows it to work easily.

### Throwable Collisions ###
Throwable objects now collide properly with the world, although they do not die. More work is needed though, to allow the different cases of collisions (side on, bottom -> top etc).