Embedded XML
### World ###
After last week's issues of loading the XML file from local directories; it has now been embedded. Because of as3's issues with XML embedding, it must be brought in as a stream of bytes and converted before loading.
The good news, is it does guarantee the order of loading, unlike the event onLoad method - so it will initialise at the point of loadLevel();
The program must be recompiled if the XML is changed.