Scaling Test
Added a new experimental peice of code for scaling images  - that will be used in the level editor, and any other tools we come up with that need it.
The code is rushed and not documented, as it was entirely experimental. A proper documentented version will be produced closer to the time it's actually required.